Crate rsdiff

rsdiff(1).

A simple implementation of rdiff(1).

this not really a 1:1 rdiff(1) implementation, it has the same concepts like generating a signature of a given buffer and then calculating the delta between the original buffer and a modified one with onenly the signature.

patching is implemented, but not ready yet!

here is a simple examples of how it works.

Examples:

Simple:

use rsdiff::Operation;

// imagine we have this simple file contains this small text.
let original = "i saw a red fox";
// now we will make a small edit to it.
let new = "i saw a red box";
// in this example we tell it to use 4 bytes as a block size.
let ops = rsdiff::diff_with_block_size(4, original, new);
for op in &ops {
 println!("{}", op);
}
// prints
// + 12..15 = "box"
// - 15..-3
assert_eq!(
    ops,
    vec![
        // insert the word "box" starting from 12 index.
        Operation::Insert { offset: 12, buffer: b"box".to_vec() },
        // from the index 15 remove 3 bytes (backword)
        // it is like if you are remving text with backspace!
        // or you can imagine it as how many bytes I need to skip from the buffer.
        Operation::Remove { offset: 15, len: 3 },
    ],
);

Insert & Remove.

use rsdiff::Operation;

let original = "hello there, do you know rust?";
let new = "hi, do you know about rustlang?";
let ops = rsdiff::diff_with_block_size(5, original, new);

assert_eq!(
    ops,
    vec![
        // add "hi, do" from offset 0 to 6.
        Operation::Insert { offset: 0, buffer: b"hi, do".to_vec() },
        // from 6 skip 15 bytes ("hello there, ").
        Operation::Remove { offset: 6, len: 15 },
        // at 16 add "about rustlang?".
        Operation::Insert { offset: 16, buffer: b"about rustlang?".to_vec() },
        // remove 5 bytes ("rust?").
        Operation::Remove { offset: 31, len: 5 },
    ],
);

Raw API.

use rsdiff::{Signature, Delta, Operation};

// imagine we have this simple file contains this small text.
let original = "i saw a red fox";
// now we will make a small edit to it.
let new = "i saw a red box";

// lets create our signature.
let mut signature = Signature::with_block_size(4, original);
// this dose not do anything itself, we need to call `calculate`.
signature.calculate();
// now we get the indexed signature.
let indexed_signature = signature.to_indexed();
// it is ready to create the `Delta`.
let mut delta = Delta::new(indexed_signature);
// we can now calculate the diff.
delta.diff(new);
let ops = delta.into_operations();
for op in &ops {
 println!("{}", op);
}
// prints
// + 12..15 = "box"
// - 15..-3
assert_eq!(
    ops,
    vec![
        // insert the word "box" starting from 12 index.
        Operation::Insert { offset: 12, buffer: b"box".to_vec() },
        // from the index 15 remove 3 bytes (backword)
        // it is like if you are remving text with backspace!
        // or you can imagine it as how many bytes I need to skip from the buffer.
        Operation::Remove { offset: 15, len: 3 },
    ],
);

Structs

Delta
Delta between two buffers, this dose not require the original buffer, but instead it only needs the original buffer signature, from there with the new modified buffer we can calculate the operations needed to upgrade the original to match the new modified one.

IndexedSignature
A Small representation of the orignal Signature. this only holds the Hash blocks calculated using Signature::calculate.

RollingHasher
An Adler-32 checksum modification with rolling operation. it is not the same algorithm as Adler-32, but acts similarly.

Signature
A Buffer Signature.

Enums

Operation
Operation to be done to upgrade from original version of the buffer to new version.

Functions

diff
Convenience function to compute Delta between two buffers. it will handle the creation of the Signature and the Delta.

diff_with_block_size
Same as diff. but with more control over the block_size.